Resolve linter errors

+++ app/helpers/transactions_helper.rb
@@ -1,12 +1,11 @@
 module TransactionsHelper
-  def edit_btn(author,trans_id)
-    if current_user.id == author
-      link_to 'Edit', edit_transaction_path(Transaction.find(trans_id))
-    end
-    end
-    def delete_btn(author,trans_id)
-      if current_user.id == author
-        link_to 'Destroy', Transaction.find(trans_id), method: :delete, data: { confirm: 'Are you sure?' } 
-      end
-    end
+  def edit_btn(author, trans_id)
+    link_to 'Edit', edit_transaction_path(Transaction.find(trans_id)) if current_user.id == author
+  end
+
+  # rubocop:disable Style/IfUnlessModifier, Layout/LineLength
+  def delete_btn(author, trans_id)
+    link_to 'Destroy', Transaction.find(trans_id), method: :delete, data: { confirm: 'Are you sure?' } if current_user.id == author
+  end
+  #  rubocop:enable Style/IfUnlessModifier, Layout/LineLength
 end
